refactor: email 서비스 계층 분리
